Rename internal to pkg in order to comply with Go 1.4+
